+++ docs/source/reference_guides/batch_evaluators.rst
@@ -1,7 +1,7 @@
 
 .. _batch_evaluators:
 
-Batch Evaluators
+Batch evaluators
 -----------------
